them order

+++ controllers/CategoryController.cs
@@ -4,9 +4,12 @@ using System.Linq;
 using System.Threading.Tasks;
 using BanCaCanh.data;
 using BanCaCanh.dto.category;
+using BanCaCanh.extensions;
 using BanCaCanh.Interface;
 using BanCaCanh.mappers;
+using BanCaCanh.models;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -16,9 +19,13 @@ namespace BanCaCanh.controllers
     [ApiController]
     public class CategoryController : ControllerBase
     {
+
+        private readonly UserManager<AppUser> _userManager;
         private readonly ICategoryRepository _categoryRepo;
-        public CategoryController(AppDbContext context, ICategoryRepository categoryRepo)
+
+        public CategoryController(UserManager<AppUser> userManager, AppDbContext context, ICategoryRepository categoryRepo)
         {
+            _userManager = userManager;
             _categoryRepo = categoryRepo;
         }
 
@@ -45,6 +52,12 @@ namespace BanCaCanh.controllers
         [Authorize]
         public async Task<IActionResult> CreateCategory([FromBody] CreateCategoryDto categoryDto)
         {
+            var username = User.GetUsername();
+            var appUser = await _userManager.FindByNameAsync(username);
+            if (appUser == null)
+            {
+                return BadRequest(new { message = "Người dùng không tồn tại" });
+            }
             var categoryModel = categoryDto.ToCreateCategoryDto();
             await _categoryRepo.CreateAsync(categoryModel);
             return CreatedAtAction(nameof(GetById), new { id = categoryModel.Id }, categoryModel.ToCategoryDto());
@@ -54,6 +67,12 @@ namespace BanCaCanh.controllers
         [Authorize]
         public async Task<IActionResult> UpdateCategory([FromRoute] int id, [FromBody] CreateCategoryDto categoryDto)
         {
+            var username = User.GetUsername();
+            var appUser = await _userManager.FindByNameAsync(username);
+            if (appUser == null)
+            {
+                return BadRequest(new { message = "Người dùng không tồn tại" });
+            }
             var categoryModel = await _categoryRepo.UpdateAsync(id, categoryDto);
             if (categoryModel == null)
             {
@@ -67,12 +86,17 @@ namespace BanCaCanh.controllers
         [Authorize]
         public async Task<IActionResult> DeleteCategory([FromRoute] int id)
         {
+            var username = User.GetUsername();
+            var appUser = await _userManager.FindByNameAsync(username);
+            if (appUser == null)
+            {
+                return BadRequest(new { message = "Người dùng không tồn tại" });
+            }
             var categoryModel = await _categoryRepo.DeleteAsync(id);
             if (categoryModel == null)
             {
                 return NotFound();
             }
-
             return NoContent();
         }
     }
